ml_service/requirements.txt
# Serving
fastapi
uvicorn
gunicorn
SQLAlchemy==2.0.41
sqlmodel==0.0.24
PyJWT==2.8.0
pytest==8.3.5
pytest-asyncio==0.26.0
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
#useful libs
passlib==1.7.4
pillow==11.3.0
python-multipart==0.0.20
# ML - Pin your versions!
numpy==2.3.1
torch==2.7.1
torchvision==0.22.1
ultralytics==8.0.239
transformers==4.53.0
huggingface_hub>=0.18.0
pillow>=9.0.0